src/__tests__/storage/monitoring.test.ts
import { MonitoringService } from '../../services/monitoring';
import { StorageEvent, StorageMetric } from '../../types/storage';
import { jest } from '@jest/globals';

// Mock external dependencies
jest.mock('winston');
jest.mock('@logtail/node');
jest.mock('@aws-sdk/client-cloudwatch');

describe('Monitoring Service', () => {
  let monitoringService: MonitoringService;

  beforeEach(() => {
    monitoringService = MonitoringService.getInstance();
  });

  describe('Event Logging', () => {
    it('should log storage events', async () => {
      const event: StorageEvent = {
        type: 'FileUpload',
        details: {
          path: 'test.txt',
          size: 1024,
        },
        metadata: {
          provider: 'aws',
        },
      };

      await monitoringService.logStorageEvent(event);
      // Verify winston logger was called
    });

    it('should handle error events', async () => {
      const event: StorageEvent = {
        type: 'Error',
        details: {
          message: 'Upload failed',
          error: new Error('Network error'),
        },
      };

      await monitoringService.logStorageEvent(event);
      // Verify error was logged correctly
    });
  });

  describe('Metrics Recording', () => {
    it('should record metrics', async () => {
      const metric: StorageMetric = {
        name: 'StorageUsage',
        value: 1024 * 1024,
        unit: 'Bytes',
        dimensions: {
          provider: 'aws',
          region: 'us-east-1',
        },
      };

      await monitoringService.recordMetric(metric);
      // Verify CloudWatch metrics were sent
    });

    it('should handle metric recording failures', async () => {
      const metric: StorageMetric = {
        name: 'InvalidMetric',
        value: -1,
        unit: 'Count',
        dimensions: {},
      };

      await expect(monitoringService.recordMetric(metric))
        .rejects
        .toThrow();
    });
  });

  describe('Health Checks', () => {
    it('should report healthy status', async () => {
      // Mock healthy system state
      const health = await monitoringService.checkStorageHealth();
      
      expect(health.status).toBe('healthy');
      expect(health.issues).toHaveLength(0);
    });

    it('should detect disk space issues', async () => {
      // Mock low disk space
      const health = await monitoringService.checkStorageHealth();
      
      expect(health.status).toBe('degraded');
      expect(health.issues).toContain(
        expect.stringContaining('disk space')
      );
    });

    it('should detect backup issues', async () => {
      // Mock failed backups
      const health = await monitoringService.checkStorageHealth();
      
      expect(health.status).toBe('degraded');
      expect(health.issues).toContain(
        expect.stringContaining('backup failed')
      );
    });

    it('should detect sync issues', async () => {
      // Mock sync problems
      const health = await monitoringService.checkStorageHealth();
      
      expect(health.status).toBe('degraded');
      expect(health.issues).toContain(
        expect.stringContaining('synchronization')
      );
    });

    it('should aggregate multiple issues', async () => {
      // Mock multiple problems
      const health = await monitoringService.checkStorageHealth();
      
      expect(health.status).toBe('unhealthy');
      expect(health.issues.length).toBeGreaterThan(1);
    });
  });

  describe('Performance Monitoring', () => {
    it('should track upload performance', async () => {
      const event: StorageEvent = {
        type: 'FileUpload',
        details: {
          path: 'test.txt',
          size: 1024 * 1024,
          duration: 1000, // 1 second
        },
      };

      await monitoringService.logStorageEvent(event);
      
      // Verify performance metrics were recorded
      expect(/* check metrics */).toBeTruthy();
    });

    it('should track sync performance', async () => {
      const event: StorageEvent = {
        type: 'SyncCompleted',
        details: {
          filesProcessed: 100,
          duration: 5000, // 5 seconds
        },
      };

      await monitoringService.logStorageEvent(event);
      
      // Verify sync metrics were recorded
      expect(/* check metrics */).toBeTruthy();
    });
  });

  describe('Alert Thresholds', () => {
    it('should trigger alerts for critical issues', async () => {
      // Mock critical system state
      const health = await monitoringService.checkStorageHealth();
      
      expect(health.status).toBe('unhealthy');
      // Verify alerts were triggered
      expect(/* check alerts */).toBeTruthy();
    });

    it('should respect alert cooldown periods', async () => {
      // Trigger multiple alerts in succession
      await monitoringService.checkStorageHealth();
      await monitoringService.checkStorageHealth();
      
      // Verify alert throttling
      expect(/* check alert count */).toBeLessThan(2);
    });
  });
});
